Expose market price for materials for each city
As a user, one way of profiting from corporations is shipping materials from one city to another to resell there.  In practice, it tends to be at most a 20% bump in overall division revenue including selling products.  This is the primary thing I expected when Corporation API came out.  
As a developer, I can understand the desire to hide this in case you don't want users profiting this way.  So, this is merely a proposal